+++ tests/pl/test_render_points.py
@@ -30,3 +30,10 @@ class TestPoints(PlotTester, metaclass=PlotTesterMeta):
 
     def test_plot_coloring_with_cmap(self, sdata_blobs: SpatialData):
         sdata_blobs.pl.render_points(color="genes", cmap="rainbow").pl.show()
+
+    def test_plot_can_stack_render_points(self, sdata_blobs: SpatialData):
+        (
+            sdata_blobs.pl.render_points(elements="blobs_points", na_color="red", size=30)
+            .pl.render_points(elements="blobs_points", na_color="blue", size=10)
+            .pl.show()
+        )